little debugs

+++ backend/controllers/product.controller.js
@@ -54,6 +54,10 @@ export const updateProduct = async (req, res) => {
 export const deleteProduct = async (req, res) => {
     const { id } = req.params;
 
+    if (!mongoose.Types.ObjectId.isValid(id)) {
+        return res.status(404).json({ success: false, message: "Invalid Product ID" });
+    }
+
     try {
         await Product.findByIdAndDelete(id);
 
@@ -61,6 +65,6 @@ export const deleteProduct = async (req, res) => {
     } catch (error) {
         console.log("error in deleting product: ", error.message);
 
-        res.status(404).json({ success: false, message: "Product not found" });
+        res.status(500).json({ success: false, message: "Server Error" });
     }
 };
